apps/gcp_linebot_localhost_katago/localhost/requirements.txt
# FastAPI stack
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Config
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# HTTP client
httpx>=0.27.0

# Go game processing (from katago/)
sgfmill
chardet

# Google Cloud SDK
google-cloud-storage>=2.18.0
